Add a reason to/remove some `//@no-rustfix` annotations
